Wed, May 17, 2023, 2:07 PM +08:00

+++ Algorithmic ToolBox/Week4/4-3.py
@@ -19,7 +19,7 @@ def majority_element_naive(elements):
 
 
 if __name__ == '__main__':
-    input_n = int(input())
+    input_n = float(input())
     input_elements = list(map(int, input().split()))
     assert len(input_elements) == input_n
     print(majority_element_naive(input_elements))
